Skip already-deleted parents in bulk removal
If these aren't skipped then an error occurs if a parent is destroyed more than
once.

Jira-Issue: MAV-2926

+++ app/models/patient_import.rb
@@ -141,7 +141,8 @@ class PatientImport < ApplicationRecord
       parent_relationships.destroy_all
 
       parents_to_check.each do |parent|
-        if parent.parent_relationships.empty? && parent.consents.empty?
+        if !parent.destroyed? && parent.parent_relationships.empty? &&
+             parent.consents.empty?
           parent.destroy!
         end
       end
@@ -172,7 +173,8 @@ class PatientImport < ApplicationRecord
       parent_relationships_without_consents.each(&:destroy!)
 
       parents_to_check.each do |parent|
-        if parent.parent_relationships.empty? && parent.consents.empty?
+        if !parent.destroyed? && parent.parent_relationships.empty? &&
+             parent.consents.empty?
           parent.destroy!
         end
       end
